// 04-countryapp/src/app/countries/interfaces/cache-store.interface.ts
import { Region } from "./region.type";
import { Country } from "./search.interface";

export interface CacheStore {
    byCapital:  TermCountries; 
    byCountry:  TermCountries;
    byRegion:   RegionCountries;
}

export interface TermCountries {
    term:       string;
    countries:  Country[]
}

export interface RegionCountries {
    region:    Region;
    countries: Country[]
}